+++ Ya/Algebra.hs
@@ -2,3 +2,4 @@ module Ya.Algebra (module Exports) where
 
 import Ya.Algebra.Abstract as Exports
 import Ya.Algebra.Definition as Exports
+import Ya.Algebra.Operators as Exports